+++ docs/CHANGELOG.md
@@ -8,6 +8,9 @@
 * Package `graph`:
   * Added `ExecOnce` and `ExecOnceN` 
   * Added `CumSum`
+  * `ConvertDType` to the same dtype is now a no-op.
+* Package `backend`:
+  * Added `DynamicUpdateSlice`, `DynamicSlice`, `ReduceAnd` and `ReduceOr`.
 * Package `tensors`:
   * Fixed race condition `Tensor.DonateBuffer`.
   * Fixed unnecessary copying of tensor data in `Tensor.MaterializeOnDevices`